feat: Implement CourseSchedulerContent component and refactor CourseScheduler
